+++ dumux/freeflow/navierstokes/momentum/localresidual.hh
@@ -129,7 +129,7 @@ public:
         return source;
     }
 
-        /*!
+    /*!
      * \brief Evaluates the mass flux over a face of a sub control volume.
      *
      * \param problem The problem