+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Differences Between DuMu<sup>x</sup> 3.9 and DuMu<sup>x</sup> 3.8
 ### Improvements and Enhancements
 
 - __SimpleH2O__: Fixed an issue where the function vaporizationEnthalpy returned the result in the wrong unit
+- __Face Centered Velocity Reconstrution__: For freeflow evaluation using the FC Staggered discretization, a reconstrution method has been implemented to collect the full velocity vector at the face center.
+- __Darcy-Brinkman Freeflow__: There is no a convenience function to add a Brinkman term to turn the Navier-Stokes model into the Darcy-Brinkman model. With this, a single domain can contain both free flow and flow through a porous medium.
+The term is added as a source term in the problem using the new helper function `addBrinkmanTerm`.
+The function uses new spatial parameter interface implemented in the new `BrinkmanSpatialParams` class (`dumux/freeflow/spatialparams.hh`). The helper function can deal with isotropic and anisotropic permeabilites.
 
 ### Immediate interface changes not allowing/requiring a deprecation period: